Extract prefix related methods into concern trait
