+++ packages/Translation/src/Exports/Drivers/BaseExporter.php
@@ -4,6 +4,7 @@ namespace Aedart\Translation\Exports\Drivers;
 
 use Aedart\Contracts\Translation\Exports\Exporter;
 use Aedart\Support\Helpers\Translation\TranslationLoaderTrait;
+use Aedart\Translation\Exports\Drivers\Concerns;
 use Aedart\Translation\Exports\Exceptions\FailedToExportTranslations;
 use Aedart\Translation\Exports\Exceptions\InvalidLocales;
 use Aedart\Translation\Exports\Exceptions\InvalidPaths;
@@ -20,6 +21,7 @@ use Throwable;
 abstract class BaseExporter implements Exporter
 {
     use TranslationLoaderTrait;
+    use Concerns\Prefixing;
 
     /**
      * Creates a new translations exporter instance
@@ -324,73 +326,6 @@ abstract class BaseExporter implements Exporter
         }
 
         return array_unique($groups);
-    }
-
-    /**
-     * Prefix each group with given namespace
-     *
-     * @param string[] $groups
-     * @param string $namespace [optional]
-     *
-     * @return string[]
-     */
-    protected function prefixGroups(array $groups, string $namespace = ''): array
-    {
-        if (empty($namespace) || empty($groups)){
-            return $groups;
-        }
-
-        return array_map(function($group) use($namespace) {
-            return $this->prefixGroup($group, $namespace);
-        }, $groups);
-    }
-
-    /**
-     * Prefix group with given namespace
-     *
-     * @param string $group
-     * @param string $namespace [optional]
-     *
-     * @return string
-     */
-    protected function prefixGroup(string $group, string $namespace = ''): string
-    {
-        if (empty($namespace)) {
-            return $group;
-        }
-
-        $separator = $this->prefixSeparator();
-
-        return "{$namespace}{$separator}{$group}";
-    }
-
-    /**
-     * Returns a namespace prefix separator
-     *
-     * @return string
-     */
-    protected function prefixSeparator(): string
-    {
-        // Use same kind of namespacing as shown in Laravel's documentation.
-        // @see https://laravel.com/docs/10.x/packages#language-files
-        return '::';
-    }
-
-    /**
-     * Removes wildcard namespace prefix from group
-     *
-     * @param string $group
-     *
-     * @return string
-     */
-    protected function removeWildcardPrefix(string $group): string
-    {
-        $prefix = '*' . $this->prefixSeparator();
-        if (str_starts_with($group, $prefix)) {
-            return str_replace($prefix, '', $group);
-        }
-
-        return $group;
     }
 
     /**
